Add options to turn off chunks of a pipeline
When developing, it can be useful to turn off validation and summaries
so we can focus on the section of a pipeline that deal with file loading
without having to wait for long-running processes.

Now we have `lein load /path/to/file` that lets us translate and load a
feed without doing further work.